+++ exportplan/views.py
@@ -131,6 +131,7 @@ class ExportPlanAdaptationForTargetMarketView(FormContextMixin, ExportPlanSectio
         context['check_duties_link'] = helpers.get_check_duties_link(self.export_plan)
         # To do pass lanaguage from export_plan object rather then  hardcoded
         context['language'] = helpers.get_cia_world_factbook_data(country='Netherlands', key='people,languages')
+        context['target_market_documents'] = json.dumps(self.export_plan['target_market_documents'])
         return context